♻️ createPartialStoreをvuex.tsに移動
